pyproject.toml
[project]
name = "postgres-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
]
